# api-docs.json
{
    "openapi":  "3.0.0",
    "info":  {
                 "title":  "Uwazi API Documentation",
                 "version":  "1.0.0",
                 "description":  "API documentation for the Uwazi Social Justice Tracking Application"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:5000"
                    }
                ],
    "components":  {
                       "securitySchemes":  {
                                               "AdminAuth":  {
                                                                 "type":  "apiKey",
                                                                 "in":  "header",
                                                                 "name":  "X-User-Id",
                                                                 "description":  "Admin user ID used for protected admin endpoints"
                                                             }
                                           }
                   },
    "paths":  {
                  "/api/comments":  {
                                        "post":  {
                                                     "summary":  "Create a new comment on a report",
                                                     "tags":  [
                                                                  "Comments"
                                                              ],
                                                     "requestBody":  {
                                                                         "required":  true,
                                                                         "content":  {
                                                                                         "application/json":  {
                                                                                                                  "schema":  {
                                                                                                                                 "type":  "object",
                                                                                                                                 "required":  [
                                                                                                                                                  "content",
                                                                                                                                                  "reportId",
                                                                                                                                                  "authorId"
                                                                                                                                              ],
                                                                                                                                 "properties":  {
                                                                                                                                                    "content":  {
                                                                                                                                                                    "type":  "string"
                                                                                                                                                                },
                                                                                                                                                    "reportId":  {
                                                                                                                                                                     "type":  "string"
                                                                                                                                                                 },
                                                                                                                                                    "authorId":  {
                                                                                                                                                                     "type":  "string"
                                                                                                                                                                 }
                                                                                                                                                }
                                                                                                                             }
                                                                                                              }
                                                                                     }
                                                                     },
                                                     "responses":  {
                                                                       "201":  {
                                                                                   "description":  "Comment created successfully"
                                                                               },
                                                                       "400":  {
                                                                                   "description":  "Validation error"
                                                                               }
                                                                   }
                                                 }
                                    },
                  "/api/comments/report/{reportId}":  {
                                                          "get":  {
                                                                      "summary":  "Get all comments for a report",
                                                                      "tags":  [
                                                                                   "Comments"
                                                                               ],
                                                                      "parameters":  [
                                                                                         {
                                                                                             "in":  "path",
                                                                                             "name":  "reportId",
                                                                                             "required":  true,
                                                                                             "schema":  {
                                                                                                            "type":  "string"
                                                                                                        }
                                                                                         }
                                                                                     ],
                                                                      "responses":  {
                                                                                        "200":  {
                                                                                                    "description":  "List of comments for the report"
                                                                                                }
                                                                                    }
                                                                  }
                                                      },
                  "/api/comments/{id}":  {
                                             "get":  {
                                                         "summary":  "Get a single comment by ID",
                                                         "tags":  [
                                                                      "Comments"
                                                                  ],
                                                         "parameters":  [
                                                                            {
                                                                                "in":  "path",
                                                                                "name":  "id",
                                                                                "required":  true,
                                                                                "schema":  {
                                                                                               "type":  "string"
                                                                                           }
                                                                            }
                                                                        ],
                                                         "responses":  {
                                                                           "200":  {
                                                                                       "description":  "Comment details"
                                                                                   },
                                                                           "404":  {
                                                                                       "description":  "Comment not found"
                                                                                   }
                                                                       }
                                                     },
                                             "put":  {
                                                         "summary":  "Update a comment",
                                                         "tags":  [
                                                                      "Comments"
                                                                  ],
                                                         "parameters":  [
                                                                            {
                                                                                "in":  "path",
                                                                                "name":  "id",
                                                                                "required":  true,
                                                                                "schema":  {
                                                                                               "type":  "string"
                                                                                           }
                                                                            }
                                                                        ],
                                                         "requestBody":  {
                                                                             "required":  true,
                                                                             "content":  {
                                                                                             "application/json":  {
                                                                                                                      "schema":  {
                                                                                                                                     "type":  "object",
                                                                                                                                     "properties":  {
                                                                                                                                                        "content":  {
                                                                                                                                                                        "type":  "string"
                                                                                                                                                                    }
                                                                                                                                                    }
                                                                                                                                 }
                                                                                                                  }
                                                                                         }
                                                                         },
                                                         "responses":  {
                                                                           "200":  {
                                                                                       "description":  "Comment updated successfully"
                                                                                   },
                                                                           "404":  {
                                                                                       "description":  "Comment not found"
                                                                                   }
                                                                       }
                                                     },
                                             "delete":  {
                                                            "summary":  "Delete a comment",
                                                            "tags":  [
                                                                         "Comments"
                                                                     ],
                                                            "parameters":  [
                                                                               {
                                                                                   "in":  "path",
                                                                                   "name":  "id",
                                                                                   "required":  true,
                                                                                   "schema":  {
                                                                                                  "type":  "string"
                                                                                              }
                                                                               }
                                                                           ],
                                                            "responses":  {
                                                                              "200":  {
                                                                                          "description":  "Comment deleted successfully"
                                                                                      },
                                                                              "404":  {
                                                                                          "description":  "Comment not found"
                                                                                      }
                                                                          }
                                                        }
                                         },
...
              },
...
}